Update code to use size_t
The`size_t` type is used to represent the size of an object.
It is defined as the return type of the sizeof operator. As an
implication size_t is used for holding an array index.

Many of the functions that would naturally return `size_t`
also return -1 when an error has occured. To hold the error
these functions still return int or long.

Being consistant with the usage of `size_t` helps reduce the
amount of casting when compiling for systems that are not
32-bit.

This is a first pass through the code to identify locations
that `int` or `long` was used when `size_t` probably should
have been used.

This is not an exastive update.

+++ api/oc_introspection.c
@@ -27,7 +27,7 @@
 #define MAX_TAG_LENGTH 20
 
 static void
-gen_idd_tag(const char *name, int device_index, char *idd_tag)
+gen_idd_tag(const char *name, size_t device_index, char *idd_tag)
 {
   int idd_tag_len =
     snprintf(idd_tag, MAX_TAG_LENGTH, "%s_%d", name, device_index);
@@ -37,7 +37,7 @@ gen_idd_tag(const char *name, int device_index, char *idd_tag)
 }
 
 int
-get_IDD_filename(int device_index, char *filename)
+get_IDD_filename(size_t device_index, char *filename)
 {
   char idd_tag[MAX_TAG_LENGTH];
   gen_idd_tag("IDD", device_index, idd_tag);
@@ -51,7 +51,7 @@ get_IDD_filename(int device_index, char *filename)
 }
 
 void
-oc_set_introspection_file(int device, const char *filename)
+oc_set_introspection_file(size_t device, const char *filename)
 {
   char idd_tag[MAX_TAG_LENGTH];
   gen_idd_tag("IDD", device, idd_tag);
@@ -215,7 +215,7 @@ oc_core_introspection_wk_handler(oc_request_t *request,
 }
 
 void
-oc_create_introspection_resource(int device)
+oc_create_introspection_resource(size_t device)
 {
   OC_DBG("oc_introspection: Initializing introspection resource");
