+++ port/freertos/ipcontext.h
@@ -19,6 +19,7 @@
 #ifndef IPCONTEXT_H
 #define IPCONTEXT_H
 
+#include <stddef.h>
 #include <stdint.h>
 
 #ifdef OC_TCP
@@ -71,7 +72,7 @@ typedef struct ip_context_t {
   tcp_context_t tcp;
 #endif
   int terminate;
-  int device;
+  size_t device;
   int shutdown_pipe[2];
 } ip_context_t;
